Introduce a separate clock source for metrics
The number of latency measurements performed in the client library is
proportional to the number of messages that are processed


On systems where reading the system clock is expensive (i.e. systems that
have a non-TSC clock source or a TSC clock source that is not exposed to
user space via VSDO (e.g. AWS EC2))

TODO:...